requirements2.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
deprecation==2.1.0
distro==1.9.0
fastapi==0.109.0
frozenlist==1.4.1
gotrue==2.1.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.4
marshmallow==3.20.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
openai==1.11.1
orjson==3.10.0
packaging==23.2
postgrest==0.15.0
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
realtime==1.0.2
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.4
supafunc==0.3.3
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.27.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==11.0.3
yarl==1.9.4
